moved to A current limit interface
